Bump dpctl quota on dppy channel
Recently there was support of new python 3.13 added into dpctl. It requires about 120 MB extra in storage to upload the new conda and wheels packages with python 3.13.
The current quota is not able to fit all the packages with the latest dpctl version. It results that every time dpctl is tiring to upload the packages, they will be cleaned up by that script.
The PR proposes to bump the quota to consider new dpctl requirement with python 3.13 support.